# src/main/java/org/openapitools/api/ImagesApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.4.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.annotation.Generated;
import jakarta.validation.Valid;
import org.openapitools.model.Error;
import org.openapitools.model.ImageCreateDto;
import org.openapitools.model.ImageReadDto;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.context.request.NativeWebRequest;

import java.util.Optional;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2023-03-07T20:09:54.805483300+03:00[Europe/Moscow]")
@Validated
@Tag(name = "ImageController", description = "the ImageController API")
public interface ImagesApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /images : Method for created image
     *
     * @param imageCreateDto (optional)
     * @return successful operation (status code 200)
     * or something wrong (status code 400)
     */
    @Operation(
            operationId = "createImage",
            summary = "Method for created image",
            tags = {"ImageController"},
            responses = {
                    @ApiResponse(responseCode = "200", description = "successful operation", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = ImageReadDto.class))
                    }),
                    @ApiResponse(responseCode = "400", description = "something wrong", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.POST,
            value = "/images",
            produces = {"application/json"},
            consumes = {"application/json"}
    )
    default ResponseEntity<ImageReadDto> createImage(
            @Parameter(name = "ImageCreateDto", description = "") @Valid @RequestBody(required = false) ImageCreateDto imageCreateDto
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType : MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"page_id\" : 6, \"link\" : \"photo link\", \"id\" : 0, \"cover_id\" : 1 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /images/{id} : Delete image by ID
     * Method for delete image
     *
     * @param id ID for delete image (required)
     * @return successful operation (status code 200)
     * or something wrong (status code 400)
     */
    @Operation(
            operationId = "deleteImage",
            summary = "Delete image by ID",
            description = "Method for delete image",
            tags = {"ImageController"},
            responses = {
                    @ApiResponse(responseCode = "200", description = "successful operation", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = Object.class))
                    }),
                    @ApiResponse(responseCode = "400", description = "something wrong", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.DELETE,
            value = "/images/{id}",
            produces = {"application/json"}
    )
    default ResponseEntity<Object> deleteImage(
            @Parameter(name = "id", description = "ID for delete image", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}