# target/classes/openapi.yaml
openapi: 3.0.1
info:
  description: This specification is intended to describe the functionality of the
    photo album application
  title: Photo album OpenAPI specification
  version: 0.0.1
servers:
- description: Dev server
  url: http://localhost:8080
paths:
...
  /albums/{coverId}:
    delete:
      description: Method for delete album
      operationId: deleteAlbum
      parameters:
      - description: ID of the cover that needs to be deleted
        explode: false
        in: path
        name: coverId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema: {}
          description: successful operation
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: something wrong
      summary: Delete album by cover ID
      tags:
      - AlbumController
      x-accepts: application/json
      x-tags:
      - tag: AlbumController
    get:
      description: Find album by cover ID
      operationId: getAlbumByCoverId
      parameters:
      - description: "cover id , which need to find an album"
        explode: false
        in: path
        name: coverId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumReadDto'
          description: successful operation
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: something wrong
      summary: Find album by cover ID
      tags:
      - AlbumController
      x-accepts: application/json
      x-tags:
      - tag: AlbumController
  /albums:
    post:
      operationId: createAlbum
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlbumCreateDto'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumReadDto'
          description: successful operation
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: something wrong
      summary: Method for created album
      tags:
      - AlbumController
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: AlbumController
...
  /images:
    post:
      operationId: createImage
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImageCreateDto'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageReadDto'
          description: successful operation
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: something wrong
      summary: Method for created image
      tags:
      - ImageController
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: ImageController
  /images/{id}:
    delete:
      description: Method for delete image
      operationId: deleteImage
      parameters:
      - description: ID for delete image
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema: {}
          description: successful operation
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: something wrong
      summary: Delete image by ID
      tags:
      - ImageController
      x-accepts: application/json
      x-tags:
      - tag: ImageController
components:
  schemas:
...
    PageReadDto:
      example:
        color: red
        tag_people: "Andrew,Egor"
        album: null
        template_page: Horizontal image and text field below
        description: some words about page
        id: 11
        tag_place: "Germany,Berlin"
        animation: slide
      properties:
        id:
          example: 11
          format: int64
          type: integer
        description:
          example: some words about page
          type: string
        color:
          example: red
          type: string
        template_page:
          example: Horizontal image and text field below
          type: string
        tag_people:
          example: "Andrew,Egor"
          type: string
        tag_place:
          example: "Germany,Berlin"
          type: string
        animation:
          example: slide
          type: string
        album:
          $ref: '#/components/schemas/AlbumReadDto'
      type: object
    AlbumCreateDto:
      example:
        created_at: 2000-01-23T04:56:07.000+00:00
        cover_id: 0
      properties:
        created_at:
          format: date-time
          type: string
        cover_id:
          format: int64
          type: integer
      type: object
    AlbumReadDto:
      example:
        created_at: 2000-01-23T04:56:07.000+00:00
        id: 0
        page:
        - color: red
          tag_people: "Andrew,Egor"
          album: null
          template_page: Horizontal image and text field below
          description: some words about page
          id: 11
          tag_place: "Germany,Berlin"
          animation: slide
        - color: red
          tag_people: "Andrew,Egor"
          album: null
          template_page: Horizontal image and text field below
          description: some words about page
          id: 11
          tag_place: "Germany,Berlin"
          animation: slide
      properties:
        id:
          format: int64
          type: integer
        created_at:
          format: date-time
          type: string
        page:
          items:
            $ref: '#/components/schemas/PageReadDto'
          type: array
      type: object
...
    ImageCreateDto:
      example:
        page_id: 0
        link: photo link
        cover_id: 6
      properties:
        link:
          example: photo link
          type: string
        page_id:
          format: int64
          type: integer
        cover_id:
          format: int64
          type: integer
      type: object
    ImageReadDto:
      example:
        page_id: 6
        link: photo link
        id: 0
        cover_id: 1
      properties:
        id:
          format: int64
          type: integer
        link:
          example: photo link
          type: string
        page_id:
          format: int64
          type: integer
        cover_id:
          format: int64
          type: integer
      required:
      - link
      type: object
...
    Error:
      properties:
        code:
          type: integer
        message:
          type: string
      type: object